refactor: delete the runtime::app re-export barrel; import types where they live
UI files now import ProjectGroup, RightTab, WorkspaceMode & co. from
tcode_core directly; mime_from_path moves to core::attachments and
MAX_TERMINALS_PER_SESSION to core::ui. The ui→runtime surface shrinks to
the documented local affordances: HostHandle, the terminal registry
types, and RuntimeEvent.
